requirements.txt
typer>=0.12.3
PyYAML>=6.0.1
openai>=1.40.0
anthropic>=0.35.0
vosk>=0.3.45
